feat(map.jinja): load a configurable list of YAML files
Lookup the configuration for `map.jinja` from:

1. builtin default `['osarch', 'os_family', 'os', 'osfinger', 'id']`

2. `defaults.yaml`: optional define a formula specific
   `map.jinja:sources`

3. global configuration lookup `map.jinja:sources`

4. formula specific `<tplroot>:map.jinja:sources` to find which
   configuration YAML files to load.

The hard coded default is backward compatible and add the minion id at
the end of the list.

If an entry does not match a `salt['config.get']` parameter, it's used
as a literal file path (with `.yaml` added if it's missing).

Each YAML file are formatted with the following top level keys:

- `values`: contains the the parameter values

- `strategy` (optional): define the merge strategy for the function
  `salt.slsutils.merge` (aggregate, list, overwrite, recurse,
  smart). The default is `smart`.

- `merge_lists` (optional): boolean to merge lists or overwrite them

The parameters values are merged in the following order:

1. initialize default values from `defaults.yaml`

2. merge the values from the YAML file
   `<tplroot>/parameters/<config>/<configvalue>.yaml` in the order
   defined by `map.jinja:sources`

3. merge the values from `<tplroot>:lookup`

4. merge the values from `salt['config.get'](<tplroot>)`